requirements.txt
python-telegram-bot>=21.0
anthropic>=0.30.0
caldav>=1.4.0
vobject>=0.9.7
openai>=1.0.0
httpx[socks]>=0.27.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
boto3>=1.34.0
cryptography>=42.0.0
